+++ packages/analytics-js-common/src/utilities/page.ts
@@ -1,7 +1,8 @@
-const onPageLeave = (callback: (...args: unknown[]) => void) => {
+const onPageLeave = (callback: (isAccessible: boolean) => void) => {
   // To ensure the callback is only called once even if more than one events
   // are fired at once.
   let pageLeft = false;
+  let isAccessible = false;
 
   function handleOnLeave() {
     if (pageLeft) {
@@ -10,7 +11,7 @@ const onPageLeave = (callback: (...args: unknown[]) => void) => {
 
     pageLeft = true;
 
-    callback();
+    callback(isAccessible);
   }
 
   // Catches the unloading of the page (e.g., closing the tab or navigating away).
@@ -18,13 +19,29 @@ const onPageLeave = (callback: (...args: unknown[]) => void) => {
   // refreshing the page, or closing the browser tab
   // Note that 'pagehide' is not supported in IE.
   // So, this is a fallback.
-  (globalThis as typeof window).addEventListener('beforeunload', handleOnLeave);
+  (globalThis as typeof window).addEventListener('beforeunload', () => {
+    isAccessible = false;
+    handleOnLeave();
+  });
+
+  (globalThis as typeof window).addEventListener('blur', () => {
+    isAccessible = true;
+    handleOnLeave();
+  });
+
+  (globalThis as typeof window).addEventListener('focus', () => {
+    pageLeft = false;
+  });
 
   // Catches the page being hidden, including scenarios like closing the tab.
-  document.addEventListener('pagehide', handleOnLeave);
+  document.addEventListener('pagehide', () => {
+    isAccessible = document.visibilityState === 'hidden';
+    handleOnLeave();
+  });
 
   // Catches visibility changes, such as switching tabs or minimizing the browser.
   document.addEventListener('visibilitychange', () => {
+    isAccessible = true;
     if (document.visibilityState === 'hidden') {
       handleOnLeave();
     } else {