+++ packages/loading-scripts/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.0.27](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.26...@rudderstack/analytics-js-loading-scripts@3.0.27) (2024-08-28)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.7.10`
 ## [3.0.26](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.25...@rudderstack/analytics-js-loading-scripts@3.0.26) (2024-08-16)
 
 ### Dependency Updates